✨ Create useDataFetching hook
- Refactor Board component using the new hook
